+++ src/clients/graphql/graphql_client.ts
@@ -51,9 +51,10 @@ export class GraphqlClient {
     const result = await this.client.post({path, type: dataType, ...params});
 
     if ((result.body as {[key: string]: unknown}).errors) {
-      throw new ShopifyErrors.GraphqlQueryError(
-        `GraphQL query returned errors: ${JSON.stringify(result.body)}`,
-      );
+      throw new ShopifyErrors.GraphqlQueryError({
+        message: 'GraphQL query returned errors',
+        response: result.body as {[key: string]: unknown},
+      });
     }
     return result;
   }